Several fixes.

+++ code/modules/projectiles/projectile/change.dm
@@ -43,22 +43,28 @@
 		return H
 
 /obj/item/projectile/change/proc/wabbajack(var/mob/M)
-	if(isliving(M) && M.stat != DEAD)
-		if(HAS_TRANSFORMATION_MOVEMENT_HANDLER(M))
-			return
-		M.handle_pre_transformation()
-		var/choice = pick(get_random_transformation_options(M))
-		var/mob/living/new_mob = apply_transformation(M, choice)
-		if(new_mob)
+
+	if(!isliving(M) || M.stat == DEAD)
+		return
+
+	if(HAS_TRANSFORMATION_MOVEMENT_HANDLER(M))
+		return
+
+	M.handle_pre_transformation()
+	var/choice = pick(get_random_transformation_options(M))
+	var/mob/living/new_mob = apply_transformation(M, choice)
+	if(new_mob)
+		new_mob.a_intent = "hurt"
+		if(M.mind)
 			for (var/spell/S in M.mind.learned_spells)
 				new_mob.add_spell(new S.type)
-			new_mob.a_intent = "hurt"
-			if(M.mind)
-				M.mind.transfer_to(new_mob)
-			else
-				new_mob.key = M.key
-			to_chat(new_mob, "<span class='warning'>Your form morphs into that of \a [choice].</span>")
-
-			qdel(M)
+			M.mind.transfer_to(new_mob)
 		else
-			to_chat(M, "<span class='warning'>Your form morphs into that of \a [choice].</span>")
+			new_mob.key = M.key
+		if(new_mob != M && !QDELETED(M))
+			qdel(M)
+	else
+		new_mob = M
+
+	if(new_mob)
+		to_chat(new_mob, SPAN_WARNING("Your form morphs into that of \a [choice]."))
